+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/action/DisableRichTextEditor.kt
@@ -4,6 +4,9 @@ import com.atlassian.performance.tools.jiraactions.api.DISABLE_RICH_TEXT_EDITOR
 import com.atlassian.performance.tools.jiraactions.api.WebJira
 import com.atlassian.performance.tools.jiraactions.api.measure.ActionMeter
 
+/**
+ * @since 3.28.0
+ */
 class DisableRichTextEditor(
     private val jira: WebJira,
     private val meter: ActionMeter